Docker: check that the installation of Chocolatey packages succeeds
In some versions of Windows with a certain set of patches installed,
the Visual C++ Build Tools installer has been observed to fail during
initialization but still exit with zero status causing a silent failure.

This patch ensures that it won't be possible to build a Docker image
where the build tools didn't actually get installed.

Change-Id: Ibc90b476dae901651002a3619f7c1dbec38bc51b
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>