Simplify export macro definition of the QML parser
We always want QML_PARSER_EXPORT to be empty. Enforce this by adding the
necessary preprocessor macros directly in qmljsglobal_p.h.  This removes
the need to define QT_CREATOR and QML_BUILD_STATIC_LIB=1 in our project
files. Also, we failed to do it correctly and always dllimported the QML
parser's symbols.

Change-Id: Ibecb2582546385c2067bd01edaba8694aa057e51
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>