Fix problems with drive root paths on Windows
This patch ensures that paths are correctly resolved against drive
roots on Windows, for example, resolvePath("C:\\foo", "..") now results
in "C:\\" rather than "C:" which is actually a relative path. It also
fixes a related issue in the product installer where one character may
have been truncated from the middle of a path since the (absolute)
prefix was always assumed to NOT end with a slash, which is always
canonical on Unix but not necessarily on Windows in the case of drive
roots where a trailing slash is necessary to differentiate from relative
path of the current drive.

Change-Id: I6c5ab1de517bb6933c3b4aebdd22236bb71c1d37
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>