Fix race condition in rule execution / TextFile auto-closing
Suppose there's a rule script that opens a TextFile and does a lengthy
operation. If a ProcessCommand finishes that has a filterOutput script
set, then it might happen that we call QScriptEngine::execute from a
QScriptEngine::execute callstack. After the inner execute finishes we
would auto-close the TextFile, unexpectedly for the code of the outer
execute call.

Delay ProcessCommand::onProcess{Finished|Error} if the ScriptEngine is
already active to avoid this situation.

Change-Id: Ibad32732f7efe9d5dbfbe1fdc143c0ea109dbb34
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>