Fix Android autotests so that the resulting APKs actually run
We don't have a NativeActivity for these apps, so remove that. We
also don't have an Application subclass, so remove the android:name
attribute from <application>. Finally, the android:name attribute of the
<activity> must specify the activity class name (fully-qualified with
the package name or with a leading dot). Opt for the latter. Lasty,
correct the System.loadLibrary calls for the lib names actually embedded
in each respective APK, and only load libraries built for the
ABIs available on the device.

Change-Id: I2b2e73a489fcda06dcf3541964eda9cf877163e8
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>