Fix Android teapot example
The latest version of the example in ndk-samples *requires* C++11, so
we switch to gnustl. We also ensure that ANativeActivity_onCreate is
exported, otherwise it'll be optimized away. The official CMake code
in ndk-samples uses this same workaround.

The ndk_helper directory in the Android NDK is deprecated, so we use the
one from ndk-samples if we find it. This should be backwards compatible
with older versions of the sample.

Change-Id: I6929a67874a8eea33d38e79bbfae83f5f8333dbb
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>