Remove usage of deprecated apkbuilder tool
Instead, we use aapt and apksigner/jarsigner. For the most part,
ApkBuilder is simply a "dumb copy" machine. It takes the initial .ap_
(which is merely a ZIP archive) and adds the dex output and native
shared libraries to it (-z and -nf options). We now just put all
of these an a properly structured directory and tell aapt to
package it. The sole purpose of "debug mode" (-d) is to determine
whether to copy gdbserver in, but Qbs already handles this bit
by not copying it to the input directory in the first place for
release builds. Finally, apkbuilder signs the APK with the Android
debug keystore by default. We instead defer that task to the
purpose built and more flexible apksigner, or jarsigner from the JDK
as a fallback for older versions of the Android SDK Build Tools.

No functionality is lost, and this also helps pave the way for
adding aar support as this will rely heavily on aapt as well.

Note that there are now *two* invocations of aapt. One to generate
R.java (which then gets compiled to a .class file and later added
to the dex output), and one to actually package the APK from the
assets, resources, dex output, native libraries, and manifest.

Change-Id: I0fd6a2ea598856d38316a3f0c8ae15a67693ddfb
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>