Allow the qbs.Xml extension to work in JavaScript files
We must return a valid object. Also, make the constructors of
XmlDomDocument and XmlDomNode members of the returned object instead of
polluting the global scope.

[ChangeLog] XmlDomDocument and XmlDomElement are deprecated and will be
removed in the 1.12 release. Use the names Xml.DomDocument and
Xml.DomElement instead, which are also now available in JavaScript files
as well as in a QML context.

Change-Id: I82e3e33c52ac7c8fdf9ba9c27491088d643b07f7
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>