Introduce Module.priority
Add the property Module.priority to be able to define an order of module
instances. This is used to resolve conflicts in module instances.

If there is more than one module instance available, the one with the
highest numerical priority is chosen.

[ChangeLog] The property Module.priority was added to provide a
mechanism to resolve conflicts in module instances.

Task-number: QBS-61
Change-Id: I6aeae91be742bb11a61345bb890992697bdf0a23
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>