Remove ModuleLoader::m_moduleSearchPaths
This member was supposed to hold the current module search paths but was
out of date as soon as an extra search path was pushed to the
ItemReader. We already searched for modules in m_modulesSearchPaths +
m_reader->allSearchPaths() to mitigate this. This led to having the same
search multiple times in the list of module paths to consider.

Change-Id: I3946d12b1708ae62e432b4bcfd18148f22ba211d
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>