Add a PropertyListVariableExpander class
This allows us to control exactly what warning message is displayed when
the expandPlistEnvironmentVariables function encounters an unknown
variable replacement, as opposed to whether or not a hardcoded message
is shown. We use this new functionality in the Info.plist processor in
order to give a much clearer message what went wrong and where the user
might look to correct it.

Change-Id: I967ea87c837f9aa2b67bad292e392b356fee8bed
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>