Android: match default NDK behavior
Unified Headers were introduced in r14 but only became default in r15.
So enable them by default if the NDK version is >= r15. Users can
manually set the property to use them with NDK r14. Only unified headers
are supported in NDK r16+ as the legacy headers are removed.

In NDK r16, the default ABI is armv7a and armv5te is deprecated and no
longer builds by default. We can't actually check the NDK version in the
qbs module due to limitations in Qbs, so we'll just use armv7a as the
global default as it has been available and been a suitable default for
a long time now.

In NDK r17, the default STL will be libc++. Set Android.ndk.appStl to
libc++_shared by default in r17 and above.

Change-Id: If0a5a60af3ceb483d0614f88dca63c7d66d52b52
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>