Fix usage of outer within Properties blocks
In EvaluatorScriptClass we're cloning the value that's referenced by
"outer" and leave out the list of alternatives. This custom cloning code
was missing the defining item of JSSourceValue, which led to an error if
the outer item was dependent on some other property.

Now, we create a full clone and clear the necessary members. This
ensures we do not have to adjust this code if another member is added to
JSSourceValue.

Task-number: QBS-1205
Change-Id: I59596894f365fa5207490922ee60959e221102b1
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>