Add qbs.targetPlatform and qbs.hostPlatform properties
This allows users to set the target platform from within project files,
since it is a scalar value. The values of qbs.targetOS and qbs.hostOS
are now derived from these new scalar properties.

[ChangeLog] Added qbs.targetPlatform and qbs.hostPlatform properties
which are scalar versions of qbs.targetOS and qbs.hostOS.
qbs.targetPlatform is a "write-only" property that can be used to
set the OS/platform that is being targeted, while qbs.targetOS
and qbs.hostOS should continue to be used to *read* the OS/platform
that is being targeted. qbs.targetOS is also now read-only.

Task-number: QBS-1070
Change-Id: Id3fd206a5b2ebee2d626f3e5ab02ecd1dc90d90a
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>