ProjectResolver: Sanitize error handling
There was lots of ad-hoc code dealing with errors, introduced by
different patches dealing with concrete bugs in an isolated manner.
Instead, we now use a structured approach that should reliably satisfy
the constraints of both command-line and IDE mode, handling errors only
on the project, product and group level.
The most visible effect of this is that we no longer run the risk of
emitting errors from disabled products, e.g. while evaluating properties
or file lists.

Change-Id: I0a27ff9d3f1f398d429c22d3ca2628b43f2e7879
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>