Make the teapot test a little less strict
Make 'unexpected files encountered' only a warning instead of an error;
it's possible and has historical precedent in the Android samples.

Further, remove th ic_launcher.png files from the list of required files
of the teapot example. These became mipmaps instead of drawables at some
point and it is not worth the effort to try to create a proper
conditional to select the right list of files.

Change-Id: If71bc394e64c19448c152b905ecff29dd436d190
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>