corelib: Gather string constants in central place
The same string literals appeared over and over again in the code base,
causing redundancy in the sources as well as at run-time. We now give
them a name and make sure they get instantiated at most once.
String literals that occur only once are converted from QLatin1String to
QStringLiteral unless they appear only in contexts that make use of
QLatin1String overloads or they are clearly outside of any hot code
paths.
This seems to result in small, but measurable performance improvements,
even if we assume the 1% changes to be noise:

========== Performance data for Resolving ==========
    Old instruction count: 3266514138
    New instruction count: 3209355927
    Relative change: -2 %
    Old peak memory usage: 29649896 Bytes
    New peak memory usage: 29436264 Bytes
    Relative change: -1 %
========== Performance data for Rule Execution ==========
    Old instruction count: 3367804753
    New instruction count: 3319029596
    Relative change: -2 %
    Old peak memory usage: 19577760 Bytes
    New peak memory usage: 19091328 Bytes
    Relative change: -3 %
========== Performance data for Null Build ==========
    Old instruction count: 608946535
    New instruction count: 604566001
    Relative change: -1 %
    Old peak memory usage: 14606384 Bytes
    New peak memory usage: 14579936 Bytes
    Relative change: -1 %

Change-Id: Ia055a52e0a4b6fe6fd0f1e7ba4bfa583cba1b0ef
Reviewed-by: Jake Petroules <jake.petroules@qt.io>