Allow "conflicting" artifacts referencing the same file, to be installed
Installing artifact A to the same location from two different products
(or multiple instances of a multiplexed product) is currently an error,
"Cannot install files X and Y to the same location..."

We only want this error message when installing artifacts pointing to
*different* file paths, to the same location (since the source is
ambiguous). We do NOT want it when installing different artifacts
pointing to the same file, to the same location, since the outcome
cannot be affected. This reduces unnecessary noise: for example, when
installing headers from a multiplexed product, the user does not need to
do extra work to ensure the files are installed by only one of the
instances.

Change-Id: Ia7a81faef5b3f27a2394b0218285c03ef3c889a3
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>