Re-use VariantValue objects
These objects are self-contained, i.e. they have no additional context
such as a scope. Therefore, there is no need to have more than one
VariantValue for the same QVariant.
We now share VariantValue objects for true, false and default QVariant,
which is easy to do, has no trade-offs such as requiring a map and is
enough to bring down the number of VariantValue objects by over 50% for
larger projects.
The overall memory savings are nothing to write home about (because
there are far more JSSourceValues), but doing fewer allocations cannot
be a bad thing.

Change-Id: Ic0c1a5cbcf3bfd8e1ee06fa71899002e899f1202
Reviewed-by: Jake Petroules <jake.petroules@qt.io>