Move the Qt library installation out of the archive product
This allows Qt to be installed as part of Qbs by default instead of only
when the archive package is built. Currently, this is enabled by default
only for installations of Qt on macOS which are relocatable (have rpath
in QT_CONFIG).

This way it will not interfere with package managers like
Homebrew/MacPorts which do not use rpath - for example, when building
against a version of Qt installed by a package manager (which is typically
built with absolute sonames), you're assumed to be building software to
be packaged by that package manager, and therefore you want to rely on
Qt as a "system" package and not copy it locally.

It can also be explicitly switched off on the command line using
products.qbsqt.deployQt:false

This also refactors the logic a bit.

Change-Id: I70b46d8fac08d34520ceab18405fa8a672158711
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>