Account for the Qbs docs build now requiring additional dependencies
Document it, and install the dependencies in the Docker images.

For the man page build, the installed version of the qbs app is now used
to get the help text rather than the non-installed version, since the
non-installed version may not be able to find all library dependencies
(like Qt) on some platforms by default. The build also now depends on
the other executables as well since their presence affects the output of
qbs --help.

Change-Id: Ic9896231aa76a346180bf30f7b2483391f5e4ea0
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>