macOS: set a compatibility version on our shared libraries
Qbs currently promises BC for the same minor release. Therefore, set the
compatibility version to the same as the SONAME version. In addition to
the mere declaration for correctness, this also helps to allow the OS
to prevent loading a mismatched qbs binary and libraries if such a
situation happens to occur.

Change-Id: Ic1cfdcab29e1efd6c51872d650d77f2ea7f2bbf1
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>