Fix spurious "unknown property" error on loading module candidates
When loading a module file, we apply property overrides from profiles
and throw an error if ther profile sets a non-existing property on the
module.
This should only be done if that module file actually gets chosen to
represent the module. In the past, it was enough to check the item
condition, but this is no longer enough, because since commit 885bd2ec92
even module files with a matching condition might not get chosen because
another one has a higher priority.
This patch fixes the problem by delaying the check until we know the
chosen candidate and applying it only on that one.
We also improve the error message: There was no context given, so users
had no clue where the problem came from.

Change-Id: I9cb1b8ab118072e6be612c466552d9dd97796e8a
Reviewed-by: Jake Petroules <jake.petroules@qt.io>