cpp module: Init all suffix and prefix properties to the empty string
...rather than leaving them undefined. This way, the string "undefined"
will not end up in file names.
It's safer to do it this way compared to checking for "undefined" in all
the places that could make use of the properties. Due to the semantics
of these properties, there is also no value in treating "undefined" in a
special way.

Task-number: QBS-1265
Change-Id: I6ce7626485baa284a6c782b513740f9d150e18c3
Reviewed-by: Jake Petroules <jake.petroules@qt.io>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>