Fix qtquickcompiler support properly
This partially reverts 193f3f02aa. The findOutput() function was really
supposed to take a file name. The fact that file paths could come in was
due to a subtle mis-use of JavaScript's String.replace() function, of
which you (absurdly) have to use the regex variant if you want to
replace more than one occurrence.

Task-number: QBS-1261
Change-Id: I8aa90e2c008c1acd7c4a9e445954d7a4d8f2234a
Reviewed-by: Jake Petroules <jake.petroules@qt.io>