Process commands: Fix environment merging for some path list variables
The value of Command.environment is supposed to be added to the general
build environment. Variables containing lists should be merged rather
than overwritten, as the code has done so far.
This is difficult to get right in the general case, as we'd need a way
to find out which variables contain lists. For now, use a fixed set of
well-known and important path list variables (PATH is particularly
relevant).

Change-Id: I860302550b8005cbe10ee7d06dcca2b5edfaf475
Reviewed-by: Jake Petroules <jake.petroules@qt.io>