Improve the documentation for cpp.separateDebugInformation
This adds information about what types of files this property causes Qbs
to generate, in order to help user discovery of the property, for
example by searching for "pdb" or "dSYM".

This change also corrects the default value listed for
cpp.separateDebugInformation, by mentioning that it also defaults to
true on Apple platforms.

Change-Id: Ida51a55a236e823b56cc05e2d8d528614d2a5dda
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>