Mark values that were set up as the defaults of built-in item properties
There are several contexts in which we need to find out whether a
property value of a built-in item is the default one that was created by
us when initializing the item.
Rather than using heuristics for that, we now set a flag on such values,
so we can always get the correct information.

Change-Id: Ieb0d87a423d6c6836a2c1d60ccf503fa0b4f801b
Reviewed-by: Jake Petroules <jake.petroules@qt.io>