setupRunEnvironment: Heuristic for Windows
On Windows, we typically link against an import library rather than the
actual dll. The dll location is not known to us, but the file is often
in a "bin" directory alongside a "lib" directory that contains the
import library. In such a case, add the "bin" directory to the PATH.

Change-Id: I21103a7ca6267d1d80c2911a19144e583c2eba95
Reviewed-by: Jake Petroules <jake.petroules@qt.io>