Fix id pointer of base items
In the following example the name of the derived product should be
"derived", but it was "base".

---BaseProduct.qbs---
Product {
    id: baseProduct
    property string foo: "base"
    name: baseProduct.foo
}

--derived.qbs--
BaseProduct {
    foo: "derived"
}

The id baseProduct still pointed to the Product item originally imported
from BaseProduct.qbs.

Fix the item pointers in the base id scopes when handling item
inheritance.

Task-number: QBS-1016
Task-number: QBS-1262
Change-Id: Id11b98aa6780f8bbbda86fda5a7d11a0e006d889
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>