Clean up the list of file dependencies before building
After adding an external file dependency, we would never remove it from
the global list again, meaning the look-up table would contain stale
entries if an external header file was moved around. Because we still
checked the status of the artifact-local list of dependencies, the
result was that in such a situation, the files including such a header
would get recompiled on every build.

[ChangeLog] Fixed constant rebuilding after moving an external header
file.

Task-number: QBS-1285
Change-Id: Id764da7485dd540f2ff64bcd890e43723f18e6cd
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>