Fix SOFT ASSERT for parametrized dependencies in Export/Module items
The SOFT ASSERT was triggered, because the Depends items in
Export/Module had the type ModuleInstance instead of ModuleParameters.

Move the assignment of that item type to ItemReaderASTVisitor instead of
adjusting it in the ModuleLoader, because this ensures that we don't
have to traverse prototypes in the adjustment procedure.

Task-number: QBS-1287
Change-Id: Ie1a1080d40b8bebe5c205f8b34298469cde5bd46
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>