Remove most load() and store() methods
For all serializable types whose load() and store() methods simply
forward these calls to their members, the methods are removed.
Instead, the serialization is now done by the persistence facility
itself, which detects such classes by the presence of a member template
with a fixed name (using SFINAE). If the template is present, it is called,
otherwise we fall back to calling load() and store() methods.

Change-Id: I42dbbdd622e0ea0db28325acc506e41c472e0f0c
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>