Save the build data of disabled products
When a product got disabled, we used to simply throw its build data
away, arguing that it was not needed anymore. However, in IDE mode, it
can easily happen that a lot of products get temporarily disabled due to
e.g. some syntax error in a central module, and then all of a sudden the
whole build is gone.
With this patch, we rescue artifact data of newly disabled products for
the case that they get re-enabled in the future.

Change-Id: I8927d808ab62d8fea4ab5691891b6fbcce9a7977
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>