Run autotests using the qbs from the install root if available
...as opposed to relying on the qbs binary existing in the same
directory as the test binaries and being usable from that location. This
is necessary for some generators (as they generate projects which call
the same qbs binary as they were generated with) where an invocation on
the qbs binary in the non-installed location may fail to find dependent
libraries when invoked from an outside environment.

Change-Id: I4d35a6f70df3ad4c48b202f73399ffe1704f78ca
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>
Reviewed-by: Jake Petroules <jake.petroules@qt.io>