Guard QScriptEngine's creation/destruction with a mutex
When building multiple configurations in one go we create one
QScriptEngine per configuration, each in its own thread, but nearly at
the same time.

It turns out that QScriptEngine's creation/destruction is not
thread-safe, which leads to all kinds of interesting effects in the
multiple configurations case.

We fix this by guarding creation/destruction of QScriptEngine with a
mutex.

Task-number: QBS-1308
Change-Id: Ie01733d5943f1fb89e1d25344c4ffa998095e394
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>