Add product name to command descriptions
The pure file name that the description usually contains does not always
provide enough context: It's not necessarily clear from a file name
which product it belongs to, and some files can be built as part of more
than one product.
Because we print the full product name, including the multiplexing
information if necessary, we also remove the JS code that provided this
kind of disambiguation for GCC compiler and linker rules only.

[ChangeLog] Command descriptions now also contain the product name.

Change-Id: I9702a6cff3faffb086e5c93e02fffa6e65fb1043
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>