Fix and rename the excludedAuxiliaryInputs property
This property did not only exclude auxiliary inputs, but all kinds of
inputs ("inputs", "auxiliaryInputs", "explicitlyDependsOn"), which
happens to be what we want, as it does not appear to make any sense to
exclude only tags from one of the relevant "input-like" properties.
But the behavior was inconsistent: When building the (product-local)
rule graph, the property was considered, but not when collecting
artifacts from other products during rule application. This is now
fixed.

[ChangeLog] The excludedAuxiliaryInputs property of the Rule item has
been renamed to excludedInputs. The old name is deprecated.

Change-Id: Id7eae3461a6e85f53b3d47750d296c3e47a18a95
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>