Do not needlessly run rules without inputs
When we dropped the requirement that rules need inputs, we did not
properly adapt the up-to-date check in the rule node. As a result, the
prepare scripts of such rules were always executed. And with commit
8e39638264, any project containing such rules would always store its
(unchanged) build graph to disk again when doing a null build.
Both problems are now fixed by doing the appropriate checks.

Change-Id: I2c910cf2b2148bfc62730df7d84ba7417205e6b9
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>