Make sure probe results are also cached for shadow products
Because we don't turn shadow products into actual products in the
ProjectResolver, the results of probes run in their context did not end
up in the build graph and the configure scripts were therefore
needlessly re-executed on the next project resolving.
Fix this by storing these probe results along with the project-level
ones.

Change-Id: I647bbedbbe3fa6f36b536fd1b80fd321894362f8
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>