Add new property explicitlyDependsOnFromDependencies
This change adds new Rule item property called
explicitlyDependsOnFromDependencies.

It allows specifying a list of file tags to consider in dependent
target artifacts which should be used as dependencies.

[ChangeLog] The Rule item property explicitlyDependsOn no longer
considers target artifacts of products that the processed product
depends on. The new property explicitlyDependsOnFromDependencies can
be used for that purpose.

Change-Id: I47a80699fe881d4075b292ad164dd6e776049a83
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>