RulesApplicator: Check for more transformer changes
Namely, invalidate the output artifact if the set of inputs or
explicitlyDependsOn has changed.
While it is theoretically possible that a rule ignores its inputs
artifacts, there is little point in catering to this unlikely case. If
we did that, we'd need extra code and data to track accesses to the
inputs map, which would cause unneeded overhead.

Change-Id: Ic63388eaa27ffcd10509e99314a3ebe65f93790e
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>