Do not needlessly throw away output artifacts
The old code in RuleNode::apply() removed an output artifact if an input
disappeared. For multiplex rules, this is almost never the right thing
to do, as there are typically multiple inputs and the artifact would get
re-generated and re-connected right away in the RulesApplicator.
Instead, remove the output artifacts only if the rule is no longer
applicable.

Change-Id: I82d4174ea2cdbaa88ecc61a34e023ca072b7c2c3
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>