RuleNode: Also consider removed inputs that no longer exist
If an artifact goes away, it is removed from the "old inputs" list of
rule nodes, because obviously we cannot reference invalid pointers
there. However, we do need to remember that there was something there,
because otherwise we will not re-apply the rule even though we have to.
For this purpose, we replace the disappeared artifact with a dummy. This
one will never appear as an input and thus its presence will cause re-
application of the rule.

Change-Id: I060db19b85ab5749ba768a7dd0a37b588cde63a7
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>