Do not generally re-apply rules just because inputs have changed
Only a small minority of rules determines their outputs based on the
content of their inputs. In order to do that, they need to do file I/O
in their outputArtifacts scripts, for instance via qbs.TextFile or
qbs.Process. All rules that don't do that do not need to get re-applied
if an input changes.

Change-Id: Icc713a54de4f60b3ef5dca523da5bfe42cc7e176
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>