Fix change tracking edge case for when no outputs get created
Our change tracking information is stored per transformer. However, if
Rule.outputArtifacts returns an empty array, then no transformer gets
created. For this case, we need to "manually" force re-application of
the rule if properties of the input artifact have changed, in case
they've been accessed in the outputArtifacts script.

Change-Id: I0f2b8a9834f5268a282c775a945e240d3d1d9763
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>