RulesApplicator: Fix assertion
There can be input artifacts that are produced by a rule that is not
connected to the current rule. This happens if at least one of the
outputs of the rule that produces the input matches the excluded inputs
of the current rule, in which case there is no connection between the
rules.

Change-Id: If51344e95e971ae89745ce7c20c2623774af42ae
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>