Fix race condition when removing empty directories
After removing generated artifacts, we check whether the respective
parent directory is now empty, and if it is, we remove it to keep the
build directory tidy.
However, we cannot do that right away as part of rule application,
because concurrently running processes of other rules might access that
directory. Therefore, we delay the directory removal step until the
Executor has finished.

Change-Id: If996e231aa34bc487bcd179d8fc7f49678519b00
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>